Refactor the Parser.LoadTable method
This change is refactoring only (no functional changes).

It encapsulates the EGT file loading, by removing the Parser.LoadTable
method into a new EGT class.

- Create new class named "EGT" in the existing Tables\EGT.cs
- Previous contents of EGT.cs (including EGTReader) become nested/hidden
types of the new EGT class
- Data loaded from the EGT file are moved: they are now private member
data of the new EGT class instead of the Parser class.
- The Parser.LoadTables method is moved: it is now the constructor of
the new EGT class

+++ GOLDEngine/Tables/FAEdge.cs
@@ -26,9 +26,9 @@ namespace GOLDEngine.Tables
         //Characters to advance on	
         public CharacterSet Characters;
         //FAState
-        public int Target;
+        public short Target;
 
-        public FAEdge(CharacterSet CharSet, int Target)
+        public FAEdge(CharacterSet CharSet, short Target)
         {
             this.Characters = CharSet;
             this.Target = Target;
